moved buttons to sidebar, moved formatLabel to utils
