+++ web/components/Canvas/Board/Board.jsx
@@ -7,6 +7,7 @@ import {
   Select,
   Chip,
 } from "@mui/material";
+import formatLabel from "@/components//helpers/utils";
 import { GameContext } from "@/components/helpers/GameContext";
 import styles from "./Board.module.scss";
 
@@ -27,27 +28,6 @@ function Board() {
     "mrs_peacock",
   ];
 
-  const rooms = [
-    "study",
-    "hall",
-    "lounge",
-    "library",
-    "billiard",
-    "dining",
-    "conservatory",
-    "ballroom",
-    "kitchen",
-  ];
-
-  const weapons = [
-    "rope",
-    "lead_pipe",
-    "knife",
-    "wrench",
-    "candlestick",
-    "revolver",
-  ];
-
   const clientSuspect = (gameState?.assignments || {})[`${clientId}`];
   const [action, setAction] = useState("");
   const [accusation, currentAccusation] = useState({});
@@ -55,16 +35,6 @@ function Board() {
   // this is a action history for the incremental demo
   const [history, setHistory] = useState([]);
 
-  // todo clean these all up into a accusation and suggestion objects. using a single object was causing async issues
-  const [accusationSuspect, setAccusationSuspect] = useState("");
-  const [accusationWeapon, setAccusationWeapon] = useState("");
-  const [accusationRoom, setAccusationRoom] = useState("");
-
-  const [suggestionSuspect, setSuggestionSuspect] = useState("");
-  const [suggestionWeapon, setSuggestionWeapon] = useState("");
-  const [suggestionRoom, setSuggestionRoom] = useState("");
-
-  const [isControlsLocked, setIsControlsLocked] = useState(true);
 
   useEffect(() => {
     if (websocket.current) {
@@ -96,22 +66,8 @@ function Board() {
       setHistory((history) => [...history, gameState?.previous_move]);
     }
 
-    // toggle controls lock
-    if (gameState?.current_turn && gameState?.assignments[clientId])
-      setIsControlsLocked(
-        gameState.current_turn != gameState.assignments[clientId]
-      );
   }, [gameState]);
 
-  function formatLabel(label) {
-    // capitalize and add a space for suspect/room/weapon labels
-    return label
-      .split(/_/g)
-      .map((word) => {
-        return word[0].toUpperCase() + word.substring(1);
-      })
-      .join(" ");
-  }
 
   return (
     <div className={styles.Board}>
@@ -153,7 +109,8 @@ function Board() {
         </div>
       )}
       <br />
-      <br />
+
+      {/* starting game */}
       {gameState.game_phase === 0 && (
         <Button
           id="start"
@@ -174,273 +131,6 @@ function Board() {
         <b>It is {formatLabel(gameState.current_turn)}&apos;s turn</b>
       )}
 
-      <br></br>
-      {/* game controls */}
-      {gameState.game_phase === 1 && (
-        <div>
-          <Button
-            id="move"
-            variant="outlined"
-            onClick={() => {
-              setAction("move");
-            }}
-            disabled={isControlsLocked}
-          >
-            Move
-          </Button>
-          <Button
-            id="suggestion"
-            variant="outlined"
-            onClick={() => {
-              setAction("suggestion");
-            }}
-            disabled={isControlsLocked}
-          >
-            Make Suggestion
-          </Button>
-          <Button
-            id="accusation"
-            variant="outlined"
-            onClick={() => {
-              setAction("accusation");
-            }}
-            disabled={isControlsLocked}
-          >
-            Make Accusation
-          </Button>
-        </div>
-      )}
-
-      {gameState.game_phase === 1 &&
-        action == "move" &&
-        (gameState?.allowed_moves || {})[clientSuspect].map((move) => {
-          return (
-            <Button
-              key="actual-move"
-              variant="outlined"
-              onClick={() => {
-                websocket?.current?.send(
-                  JSON.stringify({
-                    type: "move",
-                    suspect: clientSuspect,
-                    location: move,
-                  })
-                );
-              }}
-              disabled={isControlsLocked}
-            >
-              {move}
-            </Button>
-          );
-        })}
-
-      {/* todo lots of repeatd code beneath, clean up */}
-      {gameState.game_phase === 1 && action == "suggestion" && (
-        <div className={styles.suggestion}>
-          <div className={styles.controls}>
-            <FormControl sx={{ minWidth: 160 }}>
-              <InputLabel id="suggestion-suspect-selection-label">
-                Select Suspect
-              </InputLabel>
-              <Select
-                labelId="suggestion-suspect-selection-label"
-                id="suggestion-suspect-selection"
-                value={suggestionSuspect}
-                label="Suspect"
-                onChange={(event) => {
-                  setSuggestionSuspect(event.target.value);
-                }}
-                disabled={isControlsLocked}
-              >
-                {suspects.map((suspect) => (
-                  <MenuItem key={suspect} value={suspect}>
-                    {formatLabel(suspect)}
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-
-            <FormControl sx={{ minWidth: 160 }}>
-              <InputLabel id="suggestion-weapon-selection-label">
-                Select Weapon
-              </InputLabel>
-              <Select
-                labelId="suggestion-weapon-selection-label"
-                id="suggestion-weapon-selection"
-                value={suggestionWeapon}
-                label="Weapon"
-                onChange={(event) => {
-                  setSuggestionWeapon(event.target.value);
-                }}
-                disabled={isControlsLocked}
-              >
-                {weapons.map((weapon) => (
-                  <MenuItem key={weapon} value={weapon}>
-                    {formatLabel(weapon)}
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-
-            <FormControl sx={{ minWidth: 160 }}>
-              <InputLabel id="suggestion-room-selection-label">
-                Select Room
-              </InputLabel>
-              <Select
-                labelId="suggestion-room-selection-label"
-                id="suggestion-room-selection"
-                value={suggestionRoom}
-                label="Room"
-                onChange={(event) => {
-                  setSuggestionRoom(event.target.value);
-                }}
-                disabled={isControlsLocked}
-              >
-                {rooms.map((room) => (
-                  <MenuItem key={room} value={room}>
-                    {formatLabel(room)}
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-          </div>
-          <div className={styles.submission}>
-            <Button
-              key={`submit-suggestion`}
-              variant="contained"
-              onClick={() => {
-                websocket?.current?.send(
-                  JSON.stringify({
-                    type: "suggestion",
-                    suspect: suggestionSuspect,
-                    room: suggestionRoom,
-                    weapon: suggestionWeapon,
-                  })
-                );
-              }}
-              disabled={isControlsLocked}
-            >
-              Submit Suggestion
-            </Button>
-            <Button
-              key={`cancel-suggestion`}
-              variant="outlined"
-              onClick={() => {
-                setAccusationRoom("");
-                setSuggestionSuspect("");
-                setSuggestionWeapon("");
-              }}
-              disabled={isControlsLocked}
-            >
-              Cancel Suggestion
-            </Button>
-          </div>
-          <div className={styles.submission}></div>
-        </div>
-      )}
-
-      {gameState.game_phase === 1 && action == "accusation" && (
-        <div className={styles.accusation}>
-          <div className={styles.controls}>
-            <FormControl sx={{ minWidth: 160 }}>
-              <InputLabel id="accusation-suspect-selection-label">
-                Select Suspect
-              </InputLabel>
-              <Select
-                labelId="accusation-suspect-selection-label"
-                id="accusation-suspect-selection"
-                value={accusationSuspect}
-                label="Suspect"
-                onChange={(event) => {
-                  setAccusationSuspect(event.target.value);
-                }}
-                disabled={isControlsLocked}
-              >
-                {suspects.map((suspect) => (
-                  <MenuItem key={suspect} value={suspect}>
-                    {formatLabel(suspect)}
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-
-            <FormControl sx={{ minWidth: 160 }}>
-              <InputLabel id="accusation-weapon-selection-label">
-                Select Weapon
-              </InputLabel>
-              <Select
-                labelId="accusation-weapon-selection-label"
-                id="accusation-weapon-selection"
-                value={accusationWeapon}
-                label="Weapon"
-                onChange={(event) => {
-                  setAccusationWeapon(event.target.value);
-                }}
-                disabled={isControlsLocked}
-              >
-                {weapons.map((weapon) => (
-                  <MenuItem key={weapon} value={weapon}>
-                    {formatLabel(weapon)}
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-
-            <FormControl sx={{ minWidth: 160 }}>
-              <InputLabel id="accusation-room-selection-label">
-                Select Room
-              </InputLabel>
-              <Select
-                labelId="accusation-room-selection-label"
-                id="accusation-room-selection"
-                value={accusationRoom}
-                label="Room"
-                onChange={(event) => {
-                  setAccusationRoom(event.target.value);
-                }}
-                disabled={isControlsLocked}
-              >
-                {rooms.map((room) => (
-                  <MenuItem key={room} value={room}>
-                    {formatLabel(room)}
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-          </div>
-          <div className={styles.submission}>
-            <Button
-              key={`submit-accusation`}
-              variant="contained"
-              onClick={() => {
-                websocket?.current?.send(
-                  JSON.stringify({
-                    type: "accusation",
-                    suspect: accusationSuspect,
-                    room: accusationRoom,
-                    weapon: accusationWeapon,
-                  })
-                );
-              }}
-              disabled={isControlsLocked}
-            >
-              Submit Accusation
-            </Button>
-            <Button
-              key={`cancel-accusation`}
-              variant="outlined"
-              onClick={() => {
-                setAccusationRoom("");
-                setAccusationSuspect("");
-                setAccusationWeapon("");
-              }}
-              disabled={isControlsLocked}
-            >
-              Cancel Accusation
-            </Button>
-          </div>
-        </div>
-      )}
 
       {gameState.game_phase === 2 && (
         <div>Game Over! {formatLabel(gameState?.winner)} won!</div>
