+++ web/components/Canvas/Sidebar/Controls/index.jsx
@@ -1,0 +1,3 @@
+import Controls from "./Controls";
+
+export default Controls;
