+++ web/components/Canvas/Sidebar/Sidebar.module.scss
@@ -6,4 +6,14 @@
   align-items: center;
   height: min(100vw - 300px, 100vh);
   width: 300px;
+
+  .accusation,
+  .suggestion {
+    display: flex;
+
+    .column {
+      display: flex;
+      flex-direction: column;
+    }
+  }
 }
